Prevent double deletion of progress dialog
When `showProgress()` is passed a `100` int value, the dialog is deleted
. If the function is passed a `100` int value a second time in the same
scope, like what happens at the end of `RecalculateXIONSpent()` and
`RecalculateXIONSupply()`, it results in a segfault.

This change standardizes the process by first initializing the progress
dialog with a `0` value, updating to a maximum of a `99` value inside
the loop, and finally closing with a `100` value after the loop has
completed.

Github-Pull: #860
Rebased-From: fa0e1f301cb4c668706a14df8ae6e149c714ef6f
Signed-off-by: cevap <dev@i2pmail.org>